requirements.txt
alembic==1.15.2
annotated-types==0.6.0
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
bottleneck==1.4.2
brotli==1.0.9
certifi==2025.4.26
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.14.0
idna==3.7
mako==1.3.10
markupsafe==3.0.2
numexpr==2.10.1
pandas==2.2.3
pip==25.1
psycopg==3.2.7
psycopg2==2.9.10
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.9.1
pyjwt==2.10.1
pysocks==1.7.1
python-dateutil==2.9.0post0
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.0
sqlalchemy==2.0.40
starlette==0.46.2
typing-extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.2
wheel==0.45.1
win_inet_pton==1.1.0
scikit-learn
category_encoders
asyncio
xgboost
python-jose
bcrypt==4.0.1
passlib[bcrypt]